+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 ### Removed
 
+- `Dialog`: remove `draggable` property. `Dialog` is draggable by default, when a `title` property is set. ([@driesd](https://github.com/driesd) in [#2059](https://github.com/teamleadercrm/ui/pull/2059))
 - `Dialog`: remove `headerColor` property ([@driesd](https://github.com/driesd) in [#2058](https://github.com/teamleadercrm/ui/pull/2058))
 - `Dialog`: remove `headerIcon` property ([@driesd](https://github.com/driesd) in [#2058](https://github.com/teamleadercrm/ui/pull/2058))